Added table Booking Feature

+++ restaurant-service/src/main/java/com/dineswift/restaurant_service/payload/response/table/AvailableTimeSlot.java
@@ -16,7 +16,7 @@ public class AvailableTimeSlot {
     @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "HH:mm")
     private LocalTime endTime;
 
-    private Integer slotDurationMinutes;
+    private Long slotDurationMinutes;
 
     private Integer numberOfAvailableSeats;
